+++ controller/chatbotController.js
@@ -1,5 +1,7 @@
 const { addHistory, getHistory, deleteHistory } = require('../model/chatbotHistory');
-const fetch = require('node-fetch');
+const fetch = (...args) =>
+  import('node-fetch').then(({default: fetch}) => fetch(...args));
+
 
 // Get response message generated by chatbot
 // Used by [POST] localhost/api/chatbot/query
