+++ controller/estimatedCostController.js
@@ -24,8 +24,8 @@ const getFullCost = async (req, res) => {
     // Get ingredients price
     const ingredients_price = await getEstimatedCost.getIngredientsPrice(ingredients.id);
     
-    // Calculate ingredients price
-    const { lowPriceRequiredIngredients, highPriceRequiredIngredients } = estimateIngredientsCost(ingredients, ingredients_price);
+    // Calculate ingredients price and check out the comment under the method
+    const { lowPriceRequiredIngredients, highPriceRequiredIngredients } = getEstimatedCost.estimateIngredientsCost(ingredients, ingredients_price);
 
     if (lowPriceRequiredIngredients.length === 0 && highPriceRequiredIngredients.length === 0) {
       return res.status(404).json({
@@ -101,7 +101,7 @@ const getPartialCost = async (req, res) => {
     const ingredients_price = await getEstimatedCost.getIngredientsPrice(ingredients.id);
     
     // Calculate ingredients price
-    const { lowPriceRequiredIngredients, highPriceRequiredIngredients } = estimateIngredientsCost(ingredients, ingredients_price);
+    const { lowPriceRequiredIngredients, highPriceRequiredIngredients } = getEstimatedCost.estimateIngredientsCost(ingredients, ingredients_price);
 
     if (lowPriceRequiredIngredients.length === 0 && highPriceRequiredIngredients.length === 0) {
       return res.status(404).json({
@@ -183,94 +183,6 @@ function prepareResponseData(lowPriceRequiredIngredients, highPriceRequiredIngre
   estimatedCost.high_cost.price = estimatedCost.info.maximum_cost;
   estimatedCost.high_cost.count =  estimatedCost.high_cost.ingredients.length;
   return { estimatedCost, lowPriceID, highPriceID };
-}
-
-function estimateIngredientsCost(ingredients, ingredients_price) {
-  // Group ingredients by their id
-  var groupedIngredientsPrice = {};
-  ingredients_price.forEach(( ingredient ) => {
-    let id = ingredient.ingredient_id;
-    if (groupedIngredientsPrice[id] == undefined) {
-      groupedIngredientsPrice[id] = [];
-    }
-    groupedIngredientsPrice[id].push(ingredient);
-  })
-
-  // Find minimum purchase quantity for every ingredients
-  // Each grocery store has different price -> low total price and high total price
-  const lowPriceRequiredIngredients = [];
-  const highPriceRequiredIngredients = [];
-  if ((ingredients.id.length === ingredients.quantity.length) && (ingredients.id.length === ingredients.measurement.length)) {
-    for (let i=0; i<ingredients.id.length; i++) {
-      let target_id = ingredients.id[i];
-      let target_qty = ingredients.quantity[i];
-      let target_measurement = ingredients.measurement[i];
-
-      let ingre = groupedIngredientsPrice[target_id];
-      
-      // If target ingredient not found in the price table -> skip this ingredient
-      if (ingre) {
-        ingre = ingre.filter((item) => {
-          try {
-            let convertedUnit = convertUnits(item.unit, item.measurement, target_measurement);
-            let estimatedPurchase = 1;
-            while (convertedUnit * estimatedPurchase < target_qty) {
-              estimatedPurchase += 1;
-            }
-            item.estimation = {
-              "unit": convertedUnit,
-              "measurement": target_measurement,
-              "purchase": estimatedPurchase,
-              "total_cost": estimatedPurchase * item.price
-            }
-            return true;
-          } catch (error) {
-            return false;
-          }
-        }).map(function(item) { return item; });
-      } else {
-        ingre = [];
-      }
-
-      if (ingre.length > 0) {
-        // Find min price
-        var minIngre = ingre.reduce((prev, curr) => {
-          return prev.estimation.total_cost < curr.estimation.total_cost ? prev : curr;
-        });
-        lowPriceRequiredIngredients.push(minIngre);
-
-        // Find max price
-        var maxIngre = ingre.reduce((prev, curr) => {
-          return prev.estimation.total_cost > curr.estimation.total_cost ? prev : curr;
-        });
-        highPriceRequiredIngredients.push(maxIngre);
-      }
-    }
-  }
-
-  return {
-    lowPriceRequiredIngredients,
-    highPriceRequiredIngredients
-  };
-}
-
-function convertUnits(value, fromUnit, toUnit) {
-  const conversions = {
-    weight: { g: 1, kg: 0.001 },
-    liquid: { l: 1, ml: 1000 }
-  };
-
-  if (fromUnit === "ea" && toUnit === "ea") {
-    return value;
-  }
-
-  if (conversions.weight[fromUnit] && conversions.weight[toUnit]) {
-      return value * (conversions.weight[toUnit] / conversions.weight[fromUnit]);
-  } else if (conversions.liquid[fromUnit] && conversions.liquid[toUnit]) {
-      return value * (conversions.liquid[toUnit] / conversions.liquid[fromUnit]);
-  } else {
-      throw new Error("Invalid unit conversion");
-  }
 }
 
 module.exports = {
