+++ package.json
@@ -23,7 +23,7 @@
     "chai": "^4.5.0",
     "chai-http": "^4.4.0",
     "cors": "^2.8.5",
-    "dotenv": "^16.4.5",
+    "dotenv": "^16.6.1",
     "express": "^4.19.1",
     "express-rate-limit": "^7.5.0",
     "express-validator": "^7.2.1",
@@ -32,6 +32,7 @@
     "mocha": "^10.7.0",
     "multer": "^1.4.5-lts.1",
     "mysql2": "^3.9.2",
+    "node-fetch": "^3.3.2",
     "nutrihelp-api": "file:",
     "sinon": "^18.0.0",
     "swagger-ui-express": "^5.0.0",
